// packages/drawboard/types.ts
export interface AppState {
// ...
  editingTextElement: NonDeletedDrawboardElement | null;
// ...
  scrollX: number;
  scrollY: number;
  cursorButton: "up" | "down";
  scrolledOutside: boolean;
  name: string | null;
  isResizing: boolean;
  isRotating: boolean;
  zoom: Zoom;
// ...
  width: number;
  height: number;
  offsetTop: number;
  offsetLeft: number;
// ...
}

export type SearchMatch = {
  id: string;
  focus: boolean;
  matchedLines: {
    offsetX: number;
    offsetY: number;
    width: number;
    height: number;
    showOnCanvas: boolean;
  }[];
};

export type UIAppState = Omit<
  AppState,
  "startBoundElement" | "cursorButton" | "scrollX" | "scrollY"
>;

export type NormalizedZoomValue = number & { _brand: "normalizedZoom" };

export type Zoom = Readonly<{
  value: NormalizedZoomValue;
}>;